cleanup + desktop debug controls for demos
