+++ Assets/NanoverImd/UI/DesktopDebugUI.cs
@@ -29,6 +29,7 @@ namespace NanoverImd
         private ICollection<ServiceHub> knownServiceHubs = new List<ServiceHub>();
         private ICollection<DiscoveryEntry> knownWebSockets = new List<DiscoveryEntry>();
         private IList<string> knownSimulations = new List<string>();
+        private IList<NanoverRecordings.DemoListing> knownDemos = new List<NanoverRecordings.DemoListing>();
 
         public float interactionForceMultiplier = 1000;
 
@@ -36,6 +37,27 @@ namespace NanoverImd
         {
             GUILayout.BeginArea(new Rect(16, 16, 192, 1024));
             GUILayout.Box("Nanover iMD");
+
+            GUILayout.Box("Demos");
+
+            if (GUILayout.Button("Refresh"))
+            {
+                NanoverRecordings
+                .FetchDemosListing()
+                .AsUniTask()
+                .ContinueWith((listing) =>
+                {
+                    knownDemos = listing;
+                });
+            }
+
+            foreach (var entry in knownDemos)
+            {
+                if (GUILayout.Button(entry.Name))
+                {
+                    NanoverRecordings.LoadDemo(entry.URL).AsUniTask().ContinueWith(simulation.ConnectRecordingReader);
+                }
+            }
 
             GUILayout.Box("Connect");
