+++ Assets/NanoverImd/UI/Scene/ListDemos.cs
@@ -1,10 +1,7 @@
 ﻿using Cysharp.Threading.Tasks;
 using Nanover.Frontend.UI;
 using System.Collections.Generic;
-using System.IO;
-using System.IO.Compression;
 using UnityEngine;
-using UnityEngine.Networking;
 
 namespace NanoverImd.UI.Scene
 {
@@ -27,7 +24,6 @@ namespace NanoverImd.UI.Scene
             Refresh();
         }
 
-        [ContextMenu("Refresh Demo Listing")]
         public void Refresh()
         {
             NanoverRecordings
@@ -40,40 +36,16 @@ namespace NanoverImd.UI.Scene
             });
         }
 
-        [ContextMenu("Load First Demo")]
-        public void LoadFirstDemo()
-        {
-            LoadDemo(demos[0].URL);
-        }
-
-        public void LoadDemo(string url)
-        {
-            var dest = Path.Combine(Application.persistentDataPath, "demo.nanover.zip");
-            var index = Path.Combine(Application.persistentDataPath, "index.msgpack");
-            var messages = Path.Combine(Application.persistentDataPath, "messages.msgpack");
-
-            var request = UnityWebRequest.Get(url);
-            var handler = new DownloadHandlerFile(dest);
-            handler.removeFileOnAbort = true;
-            request.downloadHandler = handler;
-
-            var operation = request.SendWebRequest();
-
-            operation.completed += (_) =>
-            {
-                ZipFile.ExtractToDirectory(dest, Application.persistentDataPath, overwriteFiles: true);
-
-                var reader = new NanoverRecordingReader(index, messages);
-                application.Simulation.ConnectRecordingReader(reader);
-            };
-        }
-
         public void RefreshDemos()
         {
             menu.ClearChildren();
             foreach (var demo in demos)
             {
-                menu.AddItem(demo.Name, demoIcon, () => LoadDemo(demo.URL));
+                menu.AddItem(
+                    demo.Name, 
+                    demoIcon, 
+                    () => NanoverRecordings.LoadDemo(demo.URL).AsUniTask().ContinueWith(application.Simulation.ConnectRecordingReader)
+                );
             }
         }
     }
